adds donate button to appbar, removes social buttons
