+++ src/components/Analytics/index.ts
@@ -7,6 +7,7 @@ import {
   trackCopyLink,
   trackShare,
   trackVoteClick,
+  trackClick,
 } from './utils';
 
 export default PageviewTracker;
@@ -19,4 +20,5 @@ export {
   trackCopyLink,
   trackShare,
   trackVoteClick,
+  trackClick,
 };
