+++ src/components/Analytics/utils.ts
@@ -33,6 +33,7 @@ export enum EventCategory {
   EXPLORE = 'explore',
   ENGAGEMENT = 'engagement',
   VOTE_2020 = 'vote',
+  DONATE = 'donate',
 }
 
 /**
@@ -87,4 +88,8 @@ export function trackShare(label: string) {
  */
 export function trackVoteClick(label: string) {
   trackEvent(EventCategory.VOTE_2020, EventAction.CLICK_LINK, label);
+}
+
+export function trackClick(label: string) {
+  trackEvent(EventCategory.DONATE, EventAction.CLICK, label);
 }
